pyproject.toml
[project]
authors = [
  {name = "Szőke Dominik", email = "szokedominik@gmail.com"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Topic :: Software Development :: Build Tools",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Operating System :: OS Independent",
]
dependencies = [
  "requests (>=2.32.4,<3.0.0)",
  "beautifulsoup4 (>=4.13.4,<5.0.0)",
  "aiohttp (>=3.12.14,<4.0.0)",
  "tqdm (>=4.67.1,<5.0.0)",
  "mkdocstrings-python (>=1.16.12,<2.0.0)",
  "mkdocstrings (>=0.30.0,<0.31.0)",
  "asyncio (>=4.0.0,<5.0.0)",
  "pillow (>=12.0.0,<13.0.0)",
]
description = "API client for BDO market"
homepage = "https://github.com/Fizzor96/bdomarket"
license = {file = "LICENSE"}
name = "bdomarket"
readme = "README.md"
requires-python = ">=3.1,<4.0"
version = "0.3.0"

[project.urls]
Repository = "https://github.com/Fizzor96/bdomarket"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"

[project.optional-dependencies]
docs = [
  "mkdocs>=1.6.1,<2.0.0",
  "mkdocstrings[python]>=0.30.0,<1.0.0",
]

[tool.poetry.dependencies]
python = "^3.12" # replace with desired version
